+++ props-core/src/test/java/sh/props/CoordinatedAsyncTest.java
@@ -35,9 +35,9 @@ import static sh.props.source.impl.InMemory.UPDATE_REGISTRY_ON_EVERY_WRITE;
 import org.junit.jupiter.api.RepeatedTest;
 import sh.props.group.Group;
 import sh.props.source.impl.InMemory;
-import sh.props.testhelpers.AwaitAssertionTest;
-import sh.props.testhelpers.StoreAllValuesConsumer;
-import sh.props.testhelpers.TestIntProp;
+import sh.props.textfixtures.AwaitAssertionTest;
+import sh.props.textfixtures.StoreAllValuesConsumer;
+import sh.props.textfixtures.TestIntProp;
 import sh.props.tuples.Pair;
 import sh.props.tuples.Quad;
 import sh.props.tuples.Triple;
