+++ props-core/src/test/java/sh/props/CoordinatedSyncTest.java
@@ -32,8 +32,8 @@ import static sh.props.source.impl.InMemory.UPDATE_REGISTRY_ON_EVERY_WRITE;
 import org.junit.jupiter.api.Test;
 import sh.props.group.Group;
 import sh.props.source.impl.InMemory;
-import sh.props.testhelpers.AwaitAssertionTest;
-import sh.props.testhelpers.TestIntProp;
+import sh.props.textfixtures.AwaitAssertionTest;
+import sh.props.textfixtures.TestIntProp;
 import sh.props.tuples.Tuple;
 
 @SuppressWarnings("NullAway")
